// admin/src/api/generated/models/GitRefDiscoveryResponse.ts
export interface GitRefDiscoveryResponse {
    /**
     *
     * @type {string}
     * @memberof GitRefDiscoveryResponse
     */
    applicationSourceId: string;
    /**
     *
     * @type {string}
     * @memberof GitRefDiscoveryResponse
     */
    createdAt: string;
    /**
     *
     * @type {string}
     * @memberof GitRefDiscoveryResponse
     */
    errorCode?: string | null;
    /**
     *
     * @type {string}
     * @memberof GitRefDiscoveryResponse
     */
    expiresAt?: string | null;
    /**
     *
     * @type {string}
     * @memberof GitRefDiscoveryResponse
     */
    finishedAt?: string | null;
    /**
     *
     * @type {string}
     * @memberof GitRefDiscoveryResponse
     */
    id: string;
    /**
     *
     * @type {Array<GitRefResponse>}
     * @memberof GitRefDiscoveryResponse
     */
    refs: Array<GitRefResponse>;
    /**
     *
     * @type {number}
     * @memberof GitRefDiscoveryResponse
     */
    sourceVersion: number;
    /**
     *
     * @type {string}
     * @memberof GitRefDiscoveryResponse
     */
    status: string;
    /**
     *
     * @type {string}
     * @memberof GitRefDiscoveryResponse
     */
    taskId: string;
}